requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Excel export
openpyxl>=3.1.0

# Azure OpenAI
openai>=1.3.0

# OCR capabilities (for PDF text extraction fallback)
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# Google APIs (for Gmail integration)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Optional: For enhanced email capabilities
# smtplib is built-in, but these provide additional features
# sendgrid>=6.10.0
# mailgun>=0.1.1

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.0.0
